added conditional rendering and decorator for verify user and conditional views

+++ base/views.py
@@ -1,11 +1,17 @@
 # from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
 from .models import Room, Topic
 from .forms import RoomForm
+
+
+
+
 
 
 # Create Your Views Here
@@ -18,6 +24,9 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+        if request.user.is_authenticated:
+                return redirect('home')
+        
         if request.method == 'POST':
                 username = request.POST.get('username')
                 password = request.POST.get('password')
@@ -45,8 +54,6 @@ def logoutUser(request):
         return redirect('home')
 
 
-
-
 def home(request):
         q = request.GET.get('q') if request.GET.get('q') != None else ''
         rooms = Room.objects.filter(
@@ -65,7 +72,7 @@ def room(request, pk):
         context = {'room':room}
         return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
         form = RoomForm()
         if request.method == 'POST':
@@ -78,10 +85,14 @@ def createRoom(request):
         context = {'form':form}
         return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
         room = Room.objects.get(id=pk)
         form = RoomForm(instance=room)
+
+        if request.user != room.host:
+                return HttpResponse('You are not allowed here.')
+
         if request.method == 'POST':
                 form = RoomForm(request.POST, instance=room)
                 if form.is_valid():
@@ -90,9 +101,13 @@ def updateRoom(request, pk):
 
         context = {'form':form}
         return render(request, 'base/room_form.html', context)
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
         room = Room.objects.get(id=pk)
+
+        if request.user != room.host:
+                return HttpResponse('You are not allowed here.')
+        
         if request.method == 'POST':
                 room.delete()
                 return redirect('home')
